+++ bolt/create-admin.go
@@ -1,4 +1,4 @@
-package admin
+package bolt
 
 import (
 	"errors"
@@ -7,34 +7,25 @@ import (
 	"github.com/cayleygraph/cayley"
 	"github.com/cayleygraph/cayley/graph"
 	"github.com/cayleygraph/cayley/schema"
+	"github.com/oren/doc-api"
 	"golang.org/x/crypto/bcrypt"
 )
 
-// TODO: Check for duplicate email
-// TODO: Use lock to make sure between check and write we don't have one slip in
-// func CreateAdmin(h *cayley.Handle, a Admin) error {
-func (a *Admin) Create(password string) error {
-	err := validateEmail(a.Email)
+func (a *AdminService) CreateAdmin(newAdmin *admin.Admin, password string) error {
+	err := validateEmail(newAdmin.Email)
 	if err != nil {
 		return err
 	}
 
-	a.HashedPassword, err = hashPassword(password)
+	var hashedPassword string
+	hashedPassword, err = hashPassword(password)
 	if err != nil {
 		return err
 	}
 
-	// uuid := uuid.NewV1().String()
-	// t := cayley.NewTransaction()
-	// t.AddQuad(quad.Make(quad.IRI(uuid), quad.IRI("is_a"), quad.String("admin"), nil))
-	// t.AddQuad(quad.Make(quad.IRI(uuid), quad.IRI("email"), quad.String(a.Email), nil))
-	// t.AddQuad(quad.Make(quad.IRI(uuid), quad.IRI("hashed_password"), quad.String(a.HashedPassword), nil))
-	// err = store.ApplyTransaction(t)
-	err = Insert(a.Store, Admin{
-		Name:           a.Name,
-		Email:          a.Email,
-		HashedPassword: a.HashedPassword,
-	})
+	newAdmin.HashedPassword = hashedPassword
+
+	err = insert(a.Store, newAdmin)
 
 	if err != nil {
 		return err
@@ -58,7 +49,7 @@ func hashPassword(password string) (string, error) {
 	return string(bytes), err
 }
 
-func Insert(store *cayley.Handle, o interface{}) error {
+func insert(store *cayley.Handle, o interface{}) error {
 	qw := graph.NewWriter(store)
 	defer qw.Close() // don't forget to close a writer; it has some internal buffering
 	_, err := schema.WriteAsQuads(qw, o)
